requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==4.2.7
django-appconf==1.0.5
django-browser-reload==1.12.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-htmx==1.17.0
django-imagekit==5.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pilkit==3.0
Pillow==10.1.0
platformdirs==4.0.0
sqlparse==0.4.4
